requirements.txt
django==5.0.2
djangorestframework==3.14.0
flake8>=7.0.0,<8.0.0
mypy==1.8.0
pytest>=8.0.0,<9.0.0
pytest-django==4.8.0
